+++ scripts/lichessTools-init.js
@@ -50,6 +50,7 @@ lichessTools.loadTool(LiChessTools.Tools.TimelineNotify);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerEval);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerGambits);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerSettings);
+lichessTools.loadTool(LiChessTools.Tools.StickyStudySettings);
 
 lichessTools.loadTool(LiChessTools.Tools.UserTvHistory);